examples/iss-telem/pyproject.toml
[project]
name = "iss-telem-ingester"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Tacescu",email = "alextac98@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "lightstreamer-client-lib (>=2.2.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "*", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
